fix bugs in cart

+++ src/components/Cart/Cart.tsx
@@ -1,11 +1,6 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import {
-  removeProduct,
-  clearCart,
-  increaseProducts,
-  decreaseProducts,
-} from '../../redux/cartSlice';
+import { removeProduct, clearCart, decreaseProducts, addProduct } from '../../redux/cartSlice';
 import s from './Cart.module.scss';
 type Props = {};
 
@@ -14,13 +9,17 @@ const Cart: React.FC = (props: Props) => {
 
   const dispatch = useDispatch();
   const removeItem = (index: number) => {
-    dispatch(removeProduct(index));
+    if (window.confirm('Are you sure you want to remove this item?'))
+      dispatch(removeProduct(index));
+  };
+  const clearAllItems = () => {
+    if (window.confirm('Are you sure you want to clear the cart?')) dispatch(clearCart());
   };
 
   return (
     <div className="flex justify-center flex-col ">
-      <div className="text-white text-center mb-8" onClick={() => dispatch(clearCart())}>
-        очистить карзину{' '}
+      <div className="text-white text-center mb-8" onClick={clearAllItems}>
+        очистить карзину
       </div>
       <div className={s.cardContainer}>
         {state.items.map((obj: any, index: number) => {
@@ -31,16 +30,21 @@ const Cart: React.FC = (props: Props) => {
                 <h1 className="text-center">{obj.name}</h1>
               </div>
               <div className="flex justify-between gap-6 ">
-                <div className={s.minus} onClick={() => dispatch(decreaseProducts(obj.id))}>
+                <button
+                  className={s.minus}
+                  onClick={() => dispatch(decreaseProducts(obj.id))}
+                  disabled={obj.count <= 0}>
                   -
-                </div>
-                <div className="text-center">{obj.count}</div>
-                <div className={s.plus} onClick={() => dispatch(increaseProducts(obj.id))}>
+                </button>
+                <div className="text-center">{obj.count > 0 ? obj.count : 0}</div>
+                <button className={s.plus} onClick={() => dispatch(addProduct(obj))}>
                   +
-                </div>
+                </button>
               </div>
               <div className="flex gap-20 ">
-                <div className="text-center">{obj.price * obj.count}$</div>
+                <div className="text-center">
+                  {obj.price * obj.count > 0 ? obj.price * obj.count : 0}$
+                </div>
               </div>
               <div className="justify-self-end" onClick={() => removeItem(obj.id)}>
                 X
